+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* edit view: button for switching between rendered and input view for additive field values
+
 ### Changes
 
 ### Deprecated