+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - import artifical data helper and use in test
 - page to create new rule-set in editor
+- button for switching between rendered and input view for additive field values
 
 ### Changes